+++ react-native.config.js
@@ -1,5 +1,10 @@
 module.exports = {
   dependencies: {
+    'react-native-piratechain': {
+      platforms: {
+        ios: null
+      }
+    },
     'react-native-custom-tabs': {
       platforms: {
         ios: null